Update based on comments

+++ ssl/ssl_versions.cc
@@ -344,11 +344,11 @@ int SSL_CTX_set_max_proto_version(SSL_CTX *ctx, uint16_t version) {
   return set_max_version(ctx->method, &ctx->conf_max_version, version);
 }
 
-int SSL_CTX_get_min_proto_version(SSL_CTX *ctx) {
+uint16_t SSL_CTX_get_min_proto_version(SSL_CTX *ctx) {
   return ctx->conf_min_version;
 }
 
-int SSL_CTX_get_max_proto_version(SSL_CTX *ctx) {
+uint16_t SSL_CTX_get_max_proto_version(SSL_CTX *ctx) {
   return ctx->conf_max_version;
 }
